backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==1.35.3
httpx>=0.23.0,<0.28.0
python-multipart==0.0.9
python-dotenv==1.0.1
pyyaml==6.0.1
